grouping tests by classes
