--- src/test/java/com/example/feline/tests/FelineEatMeatTest.java
@@ -1,17 +1,0 @@
-package com.example.feline.tests;
-
-import com.example.Feline;
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class FelineEatMeatTest {
-
-    @Test
-    public void eatMeat() throws Exception {
-        Feline feline = new Feline();
-        var expectedTextEatMeat = "[Животные, Птицы, Рыба]";
-        var result = feline.eatMeat().toString();
-        assertEquals("Food for felines is not correct", expectedTextEatMeat, result);
-    }
-}
